split join-flow-steps into separate files; simplify birth year options
